+++ pages/_app.tsx
@@ -12,7 +12,6 @@ import * as appConfig from "../appConfig";
 import { FormInputData } from "../src/interfaces";
 import ApplicationContainer from "../src/components/ApplicationContainer";
 import { getPageTitles } from "../src/utils/utils";
-import useRouterScroll from "../src/hooks/useRouterScroll";
 import { constructProblemDetail } from "../src/utils/formDataUtils";
 import { DSProvider } from "@nypl/design-system-react-components";
 
@@ -21,7 +20,9 @@ import { useRouter } from "next/router";
 import { GetServerSideProps } from "next";
 import aaUtils from "../src/externals/aaUtils";
 import Script from "next/script";
+import { theme } from "../src/theme";
 import pkg from "../package.json";
+
 interface MyAppProps {
   Component: any;
   pageProps: any;
@@ -31,7 +32,6 @@ console.info("App Version: ", pkg.version);
 
 function MyApp({ Component, pageProps }: MyAppProps) {
   const router = useRouter();
-  useRouterScroll({ top: 640 });
   const formInitialStateCopy = { ...formInitialState };
   const formMethods = useForm<FormInputData>({ mode: "onBlur" });
   const { favIconPath, appTitle } = appConfig;
@@ -173,7 +173,7 @@ function MyApp({ Component, pageProps }: MyAppProps) {
           `,
         }}
       />
-      <DSProvider>
+      <DSProvider theme={theme}>
         <FormProvider {...formMethods}>
           <FormDataContextProvider initState={initState}>
             <ApplicationContainer problemDetail={error}>