+++ src/components/UsernameValidationFormFields/UsernameValidationFormFields.test.tsx
@@ -9,7 +9,7 @@ jest.mock("axios");
 jest.mock("react-i18next", () => {
   const en = {
     account: {
-      title: "Step 4 of 5: Customize Your Account",
+      title: "Step 4 of 5: Customize your account",
       username: {
         label: "Username",
         instruction: "5-25 alphanumeric characters. No special characters.",
@@ -21,21 +21,23 @@ jest.mock("react-i18next", () => {
           "We encourage you to select a strong password that includes: at least 8 characters, a mixture of uppercase and lowercase letters, a mixture of letters and numbers, and at least one special character <i>except</i> period (.) <br />Example: MyLib1731@<br />Password cannot contain common patterns such as consecutively repeating a character three or more times, e.g. aaaatf54 or repeating a pattern, e.g. abcabcab",
       },
       verifyPassword: {
-        label: "Verify Password",
+        label: "Verify password",
         instruction: "8-32 characters",
       },
-      showPassword: "Show Password",
+      showPassword: "Show password",
       selectLibrary: "Select a home library:",
       termsAndCondition: {
         label: "Yes, I accept the terms and conditions.",
         text: " By submitting an application, you understand and agree to our <a href='https://www.nypl.org/help/library-card/terms-conditions'>Cardholder Terms and Conditions</a> and agree to our <a href='https://www.nypl.org/help/about-nypl/legal-notices/rules-and-regulations'>Rules and Regulations</a>. To learn more about the Library’s use of personal information, please read our <a href='https://www.nypl.org/help/about-nypl/legal-notices/privacy-policy'>Privacy Policy</a>.",
       },
       errorMessage: {
-        username: "Username must be between 5-25 alphanumeric characters.",
+        username:
+          "There was a problem. Username must be between 5-25 alphanumeric characters.",
         password:
-          "Your password must be at least 8 characters, include a mixture of both uppercase and lowercase letters, include a mixture of letters and numbers, and have at least one special character except period (.)",
-        verifyPassword: "The two passwords don't match.",
-        acceptTerms: "The Terms and Conditions must be checked.",
+          "There was a problem. Your password must be at least 8 characters, include a mixture of both uppercase and lowercase letters, include a mixture of letters and numbers, and have at least one special character except period (.)",
+        verifyPassword: "There was a problem. The two passwords don't match.",
+        acceptTerms:
+          "There was a problem. The Terms and Conditions must be checked.",
       },
     },
   };
@@ -88,7 +90,7 @@ describe("UsernameValidationFormFields", () => {
     );
 
     const label = screen.getByLabelText(/Username/i);
-    const labelRequired = screen.getByText(/Required/i);
+    const labelRequired = screen.getByText(/required/i);
     const input = screen.getByRole("textbox", { name: /Username/i });
     const helperText = screen.getByText(
       "5-25 alphanumeric characters. No special characters."